n3xx: Support clock_source=external with time_source=gpsdo
Previously, the GPSDO PPS could only be selected if GPSDO was also
providing refclk. This change allows a use case of externally provided
refclk with the internal GPSDO PPS. This is useful, for example, when
wanting to use a clock with lower phase noise for sampling, but wanting
to use GPSDO PPS for coarse time sync across distribued radios. Note the
HW and FPGA designs already support this combination, and it was just be
filtered as invalid by software.

(cherry picked from commit 112f9264839c944406ad7a96bd05c308c6d637aa)